# Replace bitsConsumed with bitsLeft to optimize BitStream